+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added Pension Contribution calculation.
 - Added Tapering calculation.
 - Added Student Loan calculation.
+- Added `userSuppliedTaxCode` to calculation, only calculate tapering if this value is `false`.
 
 ## [2.9.2] - 2023-06-12Z
 - Fixed bug to ignore additional characters on valid tax code.